all without views

+++ ewm-service/src/main/java/ru/practicum/category/service/CategoryServiceImpl.java
@@ -1,6 +1,7 @@
 package ru.practicum.category.service;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import lombok.RequiredArgsConstructor;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -10,20 +11,18 @@ import ru.practicum.category.mapper.CategoryMapper;
 import ru.practicum.category.model.Category;
 import ru.practicum.category.repository.CategoryRepository;
 import ru.practicum.error.exceptions.NotFoundException;
+import ru.practicum.event.repository.EventRepository;
 
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
+@RequiredArgsConstructor
 @Transactional(readOnly = true)
 public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
-
-    @Autowired
-    public CategoryServiceImpl(CategoryRepository categoryRepository) {
-        this.categoryRepository = categoryRepository;
-    }
+    private final EventRepository eventRepository;
 
     @Transactional
     @Override
@@ -43,6 +42,9 @@ public class CategoryServiceImpl implements CategoryService {
     @Transactional
     @Override
     public void deleteCategory(Long id) {
+        if (eventRepository.existsByCategory_Id(id)) {
+            throw new DataIntegrityViolationException("Нельзя удалить категорию. Она связана с событием.");
+        }
         categoryRepository.deleteById(id);
     }
